added more stuff

+++ WoMApi/Block/TxDetail.cs
@@ -10,9 +10,9 @@
         public int Confirmations { get; set; }
         public string Txid { get; set; }
         public string Category { get; set; }
-        public double Amount { get; set; }
-        public int Amount_satoshi { get; set; }
-        public double Fee { get; set; }
-        public int Fee_satoshi { get; set; }
+        public decimal Amount { get; set; }
+        public double Amount_satoshi { get; set; }
+        public decimal Fee { get; set; }
+        public double Fee_satoshi { get; set; }
     }
 }
